Separate firmwareworldtest into a different file
