+++ src/firmware/app/primitives/BUILD
@@ -154,6 +154,14 @@ cc_test(
     name = "autochip_move_test",
     srcs = ["autochip_move_primitive_test.cpp"],
     deps = [
+        ":test_util_world",
+    ],
+)
+
+cc_library(
+    name = "test_util_world",
+    hdrs = ["test_util_world.h"],
+    deps = [
         ":autochip_move",
         ":primitive_manager",
         "@fff",
